Working out the kinks to GPU implementation

+++ printUtils.h
@@ -3,7 +3,7 @@
 
 /* Misc utilities to print vectors and matrices for debugging */
 void printVectorI (int * v, int size);
-void printVectorF (int * v, int size);
+void printVectorF (float * v, int size);
 void printMatrix(float * mat, int rows, int cols);
 void printPair(int * actions1, int * actions2, int k);
 void printSolution(float * strat1, float * strat2, int n1, int n2);
